# src/main/resources/static/swagger.yaml
openapi: 3.0.0
info:
  title: Sistema de controle de bicicletário (29/04/2023)
  description: Modelo da disciplina. Documentação relacionada ao sistema de controle de bicicletário. Toda a API está documentada neste arquivo de definição.
  contact:
    email: pasemes@uniriotec.br
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: "1"
servers:
- url: http://localhost:8020/api
  description: SwaggerHub API Auto Mocking
...
paths:
...
  /tranca/integrarNaRede:
    post:
      tags:
      - Equipamento
      summary: colocar uma tranca nova ou retornando de reparo de volta na rede de totens
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/tranca_integrarNaRede_body'
        required: true
      responses:
        "200":
          description: Dados cadastrados
        "422":
          description: Dados Inválidos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Erro'
  /tranca/retirarDaRede:
    post:
      tags:
      - Equipamento
      summary: retirar uma tranca para aposentadoria ou reparo
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/tranca_retirarDaRede_body'
        required: true
      responses:
        "200":
          description: Dados cadastrados
        "422":
          description: Dados Inválidos (ex status inválido da tranca)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Erro'
...
  /tranca:
    get:
      tags:
      - Equipamento
      summary: recupera trancas cadastradas
      responses:
        "200":
          description: 200 OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Tranca'
    post:
      tags:
      - Equipamento
      summary: Cadastrar tranca
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NovaTranca'
        required: true
      responses:
        "200":
          description: Dados cadastrados
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tranca'
        "422":
          description: Dados Inválidos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Erro'
  /tranca/{idTranca}:
    get:
      tags:
      - Equipamento
      summary: Obter tranca
      parameters:
      - name: idTranca
        in: path
        description: UUID ou número da tranca
        required: true
        style: simple
        explode: false
        schema:
          type: integer
      responses:
        "200":
          description: Tranca encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tranca'
        "404":
          description: Não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'
    put:
      tags:
      - Equipamento
      summary: Editar tranca
      parameters:
      - name: idTranca
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NovaTranca'
        required: true
      responses:
        "200":
          description: Dados cadastrados
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tranca'
        "422":
          description: Dados Inválidos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Erro'
        "404":
          description: Não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'
    delete:
      tags:
      - Equipamento
      summary: Remover tranca
      parameters:
      - name: idTranca
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: integer
      responses:
        "200":
          description: Tranca removida
        "422":
          description: Dados Inválidos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Erro'
        "404":
          description: Não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'
  /tranca/{idTranca}/bicicleta:
    get:
      tags:
      - Equipamento
      summary: Obter bicicleta na tranca
      parameters:
      - name: idTranca
        in: path
        description: UUID ou número da tranca
        required: true
        style: simple
        explode: false
        schema:
          type: integer
      responses:
        "200":
          description: Tranca encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Bicicleta'
        "404":
          description: Bicicleta não encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'
        "422":
          description: Id da tranca inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'
  /tranca/{idTranca}/trancar:
    post:
      tags:
      - Equipamento
      summary: Realiza o trancamento da tranca alterando o status da mesma de acordo. Caso receba o id da bicleta no corpo do post também altera o status da mesma e associa a tranca à bicicleta.
      parameters:
      - name: idTranca
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/idTranca_trancar_body'
        required: false
      responses:
        "200":
          description: Ação bem sucedida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tranca'
        "422":
          description: Dados Inválidos ou tranca já se encontra trancada
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Erro'
        "404":
          description: Não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'
  /tranca/{idTranca}/destrancar:
    post:
      tags:
      - Equipamento
      summary: Realiza o destrancamento da tranca alterando o status da mesma de acordo. Caso receba o id da bicleta no corpo do post também altera o status da mesma e desassocia a tranca à bicicleta.
      parameters:
      - name: idTranca
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/idTranca_destrancar_body'
        required: false
      responses:
        "200":
          description: Ação bem sucedida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tranca'
        "422":
          description: Dados Inválidos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Erro'
        "404":
          description: Não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'
  /tranca/{idTranca}/status/{acao}:
    post:
      tags:
      - Equipamento
      summary: Alterar status da tranca
      parameters:
      - name: idTranca
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: integer
      - name: acao
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
          format: "ENUM['DESTRANCAR', 'TRANCAR']"
      responses:
        "200":
          description: Ação bem sucedida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tranca'
        "422":
          description: Dados Inválidos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Erro'
        "404":
          description: Não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'
...
components:
  schemas:
...
    NovaBicicleta:
      type: object
      properties:
        marca:
          type: string
        modelo:
          type: string
        ano:
          type: string
          format: year
        numero:
          type: integer
        status:
          type: string
          format: "enum['DISPONÍVEL','EM_USO', 'NOVA', 'APOSENTADA', 'REPARO_SOLICITADO', 'EM_REPARO' ]"
    Bicicleta:
      allOf:
      - type: object
        properties:
          id:
            type: integer
      - $ref: '#/components/schemas/NovaBicicleta'
...
    NovaTranca:
      type: object
      properties:
        numero:
          minimum: 0
          exclusiveMinimum: true
          type: integer
        localizacao:
          type: string
          format: coordinates
        anoDeFabricacao:
          type: string
          format: year
        modelo:
          type: string
        status:
          type: string
          format: "enum['LIVRE', 'OCUPADA', 'NOVA', 'APOSENTADA', 'EM_REPARO']"
    Tranca:
      allOf:
      - type: object
        properties:
          id:
            type: integer
          bicicleta:
            type: integer
      - $ref: '#/components/schemas/NovaTranca'
...
    Erro:
      required:
      - codigo
      - mensagem
      type: object
      properties:
        codigo:
          type: string
        mensagem:
          type: string
...
    tranca_integrarNaRede_body:
      type: object
      properties:
        idTotem:
          type: integer
        idTranca:
          type: integer
        idFuncionario:
          type: integer
    tranca_retirarDaRede_body:
      type: object
      properties:
        idTotem:
          type: integer
        idTranca:
          type: integer
        idFuncionario:
          type: integer
        statusAcaoReparador:
          type: string
          description: "Status escolhido como destino do equipamento, uma vez que o reparo tenha sido solicitado anteriormente."
          format: "enum['APOSENTADA', 'EM_REPARO' ]"
    idTranca_trancar_body:
      type: object
      properties:
        bicicleta:
          type: integer
    idTranca_destrancar_body:
      type: object
      properties:
        bicicleta:
          type: integer
...